Cria metodos adicionarDisciplinaAoProfessor, removerProfessor, verificarProfessor, adicionarTurmasAoProfessor, remodelar getProfessores, cadastrarProfessor, editarProfessor e sobrescreverArquivo, e criar esses metodos na classe Administrador

+++ src/dao/ProfessorDAO.java
@@ -8,13 +8,14 @@ import java.util.ArrayList;
 
 import modelo.Disciplina;
 import modelo.Professor;
+import modelo.Turma;
 
 public class ProfessorDAO {
 
     private ArrayList<Professor> listaProfessores = new ArrayList<>();
 
-public ArrayList<Professor> lerArquivo() {
-
+    public ArrayList<Professor> getProfessores() {
+    listaProfessores.clear();
     try (BufferedReader leitor = new BufferedReader(new FileReader("professores.txt"))) {
         String linha;
         while ((linha = leitor.readLine()) != null) {
@@ -31,6 +32,23 @@ public ArrayList<Professor> lerArquivo() {
                     disciplinas.add(new Disciplina(nomeDisciplina));
                 }
             }
+            String turmasTexto = dados[17].replace("[", "").replace("]", "").trim();
+            String[] turmasLecionadas = turmasTexto.split(",");
+
+            ArrayList<Turma> turmas = new ArrayList<>();
+
+            for (String t : turmasLecionadas) {
+                t = t.trim();
+                if (!t.isEmpty()) {
+                    String[] partes = t.split("\\|"); // separa ano do nome
+                    if (partes.length == 2) {
+                        int ano = Integer.parseInt(partes[0].trim());
+                        String nome = partes[1].trim();
+                        turmas.add(new Turma(ano, nome));
+                    }
+                }
+            }
+
 
             Professor professor = new Professor(
                 dados[0], // nome
@@ -52,6 +70,7 @@ public ArrayList<Professor> lerArquivo() {
                 disciplinas // disciplinas leciona
             );
 
+            professor.setTurmasLeciona(turmas);
             listaProfessores.add(professor);
         }
     } catch (IOException e) {
@@ -62,59 +81,64 @@ public ArrayList<Professor> lerArquivo() {
 }
 
     public void cadastrarProfessor(Professor professor){
-        try{
-            /*nome, cpf, rg, matricula, email, dia, mes, ano, rua, bairro, cidade, numeroCasa, complemento, nTelefone, senha, formacaoAcademica, ArrayList<Disciplina>disciplinasLeciona */
-
+        if(verificarProfessor(professor.getMatricula()) == false){//verifica se o professor existe, caso não exista ele cadastra o professor
+            try{
             FileWriter escritor = new FileWriter("professores.txt", true);
-            escritor.write(professor.getNome() + ";" + professor.getCpf() + ";" + professor.getRg() + ";" + professor.getMatricula() + ";" + professor.getEmail() + ";" + professor.getDiaNasc() + ";" + professor.getMesNasc() + ";" + professor.getAnoNasc() + ";" + professor.getNomeRua() + ";" + professor.getNomeBairro() + ";" + professor.getNomeCidade() + ";" + professor.getNumeroCasa() + ";" + professor.getComplemento() + ";" + professor.getNumeroTelefone() + ";" + professor.getSenha() + ";" + professor.getFormacaoAcademica() + ";" + professor.getDisciplinasLeciona() + "\n");
+            escritor.write(professor.getNome() + ";" + professor.getCpf() + ";" + professor.getRg() + ";" + professor.getMatricula() + ";" + professor.getEmail() + ";" + professor.getDiaNasc() + ";" + professor.getMesNasc() + ";" + professor.getAnoNasc() + ";" + professor.getNomeRua() + ";" + professor.getNomeBairro() + ";" + professor.getNomeCidade() + ";" + professor.getNumeroCasa() + ";" + professor.getComplemento() + ";" + professor.getNumeroTelefone() + ";" + professor.getSenha() + ";" + professor.getFormacaoAcademica() + ";" + professor.getDisciplinasLeciona() + ";" + professor.getTurmasLeciona() + "\n");
             escritor.close();
             System.out.println("Professor adicionado com sucesso!");
             
-            // System.out.println(professor.getDisciplinasLeciona());
-        }catch(IOException e){
-            e.getMessage();
-            // System.out.println("Erro");
+            }catch(IOException e){
+                System.err.println("Erro ao cadastrar Professor " + e.getMessage());
+            }
+        }else{
+            System.out.println("Professor já cadastrado no sistema");
         }
+        
     }
 
     public void editarProfessor(Professor novoProfessor, String matricula){
-        listaProfessores = lerArquivo();
-        boolean encontrado = false;
+        if(verificarProfessor(matricula)){
+            listaProfessores = getProfessores();
+            boolean encontrado = false;
 
-        for (Professor p : listaProfessores) {
-            if (p.getMatricula().equals(matricula)) {
-                // Atualiza os dados
-                p.setNome(novoProfessor.getNome());
-                p.setCpf(novoProfessor.getCpf());
-                p.setRg(novoProfessor.getRg());
-                p.setEmail(novoProfessor.getEmail());
-                p.setDiaNasc(novoProfessor.getDiaNasc());
-                p.setMesNasc(novoProfessor.getMesNasc());
-                p.setAnoNasc(novoProfessor.getAnoNasc());
-                p.setNomeRua(novoProfessor.getNomeRua());
-                p.setNomeBairro(novoProfessor.getNomeBairro());
-                p.setNomeCidade(novoProfessor.getNomeCidade());
-                p.setNumeroCasa(novoProfessor.getNumeroCasa());
-                p.setComplemento(novoProfessor.getComplemento());
-                p.setNumeroTelefone(novoProfessor.getNumeroTelefone());
-                p.setSenha(novoProfessor.getSenha());
-                p.setFormacaoAcademica(novoProfessor.getFormacaoAcademica());
-                p.setDisciplinasLeciona(novoProfessor.getDisciplinasLeciona());
-                encontrado = true;
-                break;
+            for (Professor p : listaProfessores) {
+                if (p.getMatricula().equals(matricula)) {
+                    // Atualiza os dados
+                    p.setNome(novoProfessor.getNome());
+                    p.setCpf(novoProfessor.getCpf());
+                    p.setRg(novoProfessor.getRg());
+                    p.setEmail(novoProfessor.getEmail());
+                    p.setDiaNasc(novoProfessor.getDiaNasc());
+                    p.setMesNasc(novoProfessor.getMesNasc());
+                    p.setAnoNasc(novoProfessor.getAnoNasc());
+                    p.setNomeRua(novoProfessor.getNomeRua());
+                    p.setNomeBairro(novoProfessor.getNomeBairro());
+                    p.setNomeCidade(novoProfessor.getNomeCidade());
+                    p.setNumeroCasa(novoProfessor.getNumeroCasa());
+                    p.setComplemento(novoProfessor.getComplemento());
+                    p.setNumeroTelefone(novoProfessor.getNumeroTelefone());
+                    p.setSenha(novoProfessor.getSenha());
+                    p.setFormacaoAcademica(novoProfessor.getFormacaoAcademica());
+                    p.setDisciplinasLeciona(novoProfessor.getDisciplinasLeciona());
+                    p.setTurmasLeciona(novoProfessor.getTurmasLeciona());
+                    encontrado = true;
+                    break;
+                }
+            }
+
+            if (encontrado) {
+                sobrescreverArquivo(); // sobrescreve o arquivo atualizado
+                System.out.println("Professor atualizado com sucesso!");
+            } else {
+                System.out.println("Professor não encontrado.");
             }
         }
-
-        if (encontrado) {
-            sobrescreverArquivo(); // sobrescreve o arquivo atualizado
-            System.out.println("Professor atualizado com sucesso!");
-        } else {
-            System.out.println("Professor não encontrado.");
-        }
+        
     }
 
     public void listarProfessores(){
-        listaProfessores = lerArquivo();
+        listaProfessores = getProfessores();
         for(Professor professor : listaProfessores){
             System.out.println(professor.getNome());
         }
@@ -124,7 +148,7 @@ public ArrayList<Professor> lerArquivo() {
         try{
             FileWriter escritor = new FileWriter("professores.txt", false);
             for(Professor professor : listaProfessores){
-                escritor.write(professor.getNome() + ";" + professor.getCpf() + ";" + professor.getRg() + ";" + professor.getMatricula() + ";" + professor.getEmail() + ";" + professor.getDiaNasc() + ";" + professor.getMesNasc() + ";" + professor.getAnoNasc() + ";" + professor.getNomeRua() + ";" + professor.getNomeBairro() + ";" + professor.getNomeCidade() + ";" + professor.getNumeroCasa() + ";" + professor.getComplemento() + ";" + professor.getNumeroTelefone() + ";" + professor.getSenha() + ";" + professor.getFormacaoAcademica() + ";" + professor.getDisciplinasLeciona() + "\n");
+                escritor.write(professor.getNome() + ";" + professor.getCpf() + ";" + professor.getRg() + ";" + professor.getMatricula() + ";" + professor.getEmail() + ";" + professor.getDiaNasc() + ";" + professor.getMesNasc() + ";" + professor.getAnoNasc() + ";" + professor.getNomeRua() + ";" + professor.getNomeBairro() + ";" + professor.getNomeCidade() + ";" + professor.getNumeroCasa() + ";" + professor.getComplemento() + ";" + professor.getNumeroTelefone() + ";" + professor.getSenha() + ";" + professor.getFormacaoAcademica() + ";" + professor.getDisciplinasLeciona() + ";" + professor.getTurmasLeciona() + "\n");
             }
             System.out.println("Arquivo sobrescrito!");
             escritor.close();
@@ -134,7 +158,7 @@ public ArrayList<Professor> lerArquivo() {
     }
 
     public boolean buscarPorNome(String nome){
-        ArrayList<Professor> listaProfessores = lerArquivo();
+        ArrayList<Professor> listaProfessores = getProfessores();
         boolean resp = false;
 
         for (Professor professor : listaProfessores) {
@@ -147,7 +171,7 @@ public ArrayList<Professor> lerArquivo() {
     }
 
     public boolean buscarPorCpf(String cpf){
-        listaProfessores = lerArquivo();
+        listaProfessores = getProfessores();
         boolean resp = false;
 
         for (Professor professor : listaProfessores) {
@@ -159,7 +183,7 @@ public ArrayList<Professor> lerArquivo() {
     }
 
     public boolean buscarPorDisciplina(String nomeDisciplina, String matricula){
-        listaProfessores = lerArquivo();
+        listaProfessores = getProfessores();
         boolean resp = false;
         for(Professor professor : listaProfessores){
             if(professor.getMatricula().equals(matricula)){
@@ -174,4 +198,77 @@ public ArrayList<Professor> lerArquivo() {
         return resp;
     }
 
+    public void adicionarDisciplinaAoProfessor(Disciplina disciplina, String matriculaProfessor){
+        if(verificarProfessor(matriculaProfessor)){
+            listaProfessores = getProfessores();
+            for(Professor professor : listaProfessores){
+                if(professor.getMatricula().equals(matriculaProfessor)){
+                    //Testa se existe disciplina com mesmo nome no ArrayList
+                    for(int i=0; i<professor.getDisciplinasLeciona().size(); i++){
+                        if(professor.getDisciplinasLeciona().get(i).getNome().toUpperCase().equals(disciplina.getNome().toUpperCase())){
+                            System.out.println("Professor já tem essa disciplina cadastrada!");
+                            return;
+                        }
+                    }
+                    //só adiciona se não achar disciplina igual já no ArrayList
+                    professor.getDisciplinasLeciona().add(disciplina);
+                    System.out.println("Disciplina adicionada com sucesso!");
+                    sobrescreverArquivo();
+                    return;
+                }
+            }
+        }else{
+            System.out.println("Erro ao adicionar disciplina!");
+        }
+    }
+
+    public void removerProfessor(String matricula){
+        if(verificarProfessor(matricula)){
+            boolean resp = false;
+            listaProfessores = getProfessores();
+            Professor professorRemover = null;
+            for(Professor professor : listaProfessores){
+                if(professor.getMatricula().equals(matricula)){
+                    professorRemover = professor;
+                    resp = true;
+                    break;
+                }
+            }
+            if(resp){
+                listaProfessores.remove(professorRemover);
+                sobrescreverArquivo();
+            }
+        }
+        
+    }
+
+    public boolean verificarProfessor(String matricula){
+        listaProfessores = getProfessores();
+        boolean resp = false;
+        for(Professor professor : listaProfessores){
+            if(professor.getMatricula().equals(matricula)){
+                resp = true;
+            }
+        }
+        return resp;
+    }
+    //FALTA CONCERTAR ESSE MÉTODO DE adicionarTurmasAoProfessor!!! Ele muda a turma ao invés de adicionar no final do ArrayList
+    public void adicionarTurmasAoProfessor(Turma turma, String matriculaProfessor){
+        listaProfessores = getProfessores();
+        if(verificarProfessor(matriculaProfessor)){
+            for(Professor p : listaProfessores){
+                if(p.getMatricula().equals(matriculaProfessor)){
+                    if(p.getTurmasLeciona().contains(turma)){
+                        System.out.println("Turma já vinculada ao professor");
+                        return;
+                    }else{
+                        p.getTurmasLeciona().add(turma);
+                        sobrescreverArquivo();
+                        System.out.println("turma adicionada com sucesso");
+                        return;
+                    }
+                }
+            }
+        }
+    }
 }
